# src/api/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"SteamGame": {
				"properties": {
					"appid": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"playtime_forever": {
						"type": "number",
						"format": "double"
					},
					"img_icon_url": {
						"type": "string"
					},
					"img_logo_url": {
						"type": "string"
					},
					"playtime_windows_forever": {
						"type": "number",
						"format": "double"
					},
					"playtime_mac_forever": {
						"type": "number",
						"format": "double"
					},
					"playtime_linux_forever": {
						"type": "number",
						"format": "double"
					},
					"rtime_last_played": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"appid"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SteamGamesResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/SteamGame"
						},
						"type": "array"
					},
					"error": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SteamErrorResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					},
					"error": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SteamPlayer": {
				"properties": {
					"steamid": {
						"type": "string"
					},
					"personaname": {
						"type": "string"
					},
					"profileurl": {
						"type": "string"
					},
					"avatar": {
						"type": "string"
					},
					"avatarmedium": {
						"type": "string"
					},
					"avatarfull": {
						"type": "string"
					},
					"personastate": {
						"type": "number",
						"format": "double"
					},
					"realname": {
						"type": "string"
					},
					"loccountrycode": {
						"type": "string"
					}
				},
				"required": [
					"steamid",
					"personaname"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SteamPlayerResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/SteamPlayer"
					},
					"error": {
						"type": "string"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiError": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					},
					"error": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MessageResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "Launcher of Launchers API",
		"version": "1.0.0",
		"description": "Launcher of launchers",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/steam/{steamId}/owned-games": {
			"get": {
				"operationId": "GetOwnedGames",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SteamGamesResponse"
								}
							}
						}
					},
					"400": {
						"description": "SteamID inválido",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SteamErrorResponse"
								}
							}
						}
					},
					"503": {
						"description": "Steam API no configurada",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SteamErrorResponse"
								}
							}
						}
					}
				},
				"description": "Devuelve el listado de juegos en propiedad de un usuario de Steam.",
				"tags": [
					"Steam"
				],
				"security": [],
				"parameters": [
					{
						"description": "SteamID64 de 17 dígitos (ej: 76561198000000000)",
						"in": "path",
						"name": "steamId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Incluir nombre e imágenes de cada juego (por defecto true)",
						"in": "query",
						"name": "includeAppInfo",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/steam/{steamId}/player-summary": {
			"get": {
				"operationId": "GetPlayerSummary",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SteamPlayerResponse"
								}
							}
						}
					},
					"404": {
						"description": "Jugador no encontrado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SteamErrorResponse"
								}
							}
						}
					},
					"503": {
						"description": "Steam API no configurada",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SteamErrorResponse"
								}
							}
						}
					}
				},
				"description": "Devuelve el resumen del perfil de un usuario de Steam.",
				"tags": [
					"Steam"
				],
				"security": [],
				"parameters": [
					{
						"description": "SteamID64 de 17 dígitos",
						"in": "path",
						"name": "steamId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/steam/{steamId}/recently-played": {
			"get": {
				"operationId": "GetRecentlyPlayedGames",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SteamGamesResponse"
								}
							}
						}
					},
					"503": {
						"description": "Steam API no configurada",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SteamErrorResponse"
								}
							}
						}
					}
				},
				"description": "Devuelve los juegos jugados recientemente (últimas 2 semanas).",
				"tags": [
					"Steam"
				],
				"security": [],
				"parameters": [
					{
						"description": "SteamID64 de 17 dígitos",
						"in": "path",
						"name": "steamId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Número máximo de juegos a devolver (por defecto 10)",
						"in": "query",
						"name": "count",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/launchers/{launcherType}/callback": {
			"get": {
				"operationId": "HandleCallback",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					},
					"400": {
						"description": "Parámetros inválidos",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Callback OAuth2 para launchers que redirigen al backend directamente.\r\nRuta pública — no requiere JWT.",
				"tags": [
					"Launchers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "launcherType",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "code",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "state",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}